requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
beautifulsoup4
google-generativeai
argparse